backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic-settings==2.7.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
playwright==1.49.1
bleach==6.2.0
typer==0.15.1
python-slugify==8.0.4
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
